It kinda do
Got this to where you can take a CSV file in, reorder the columns with Format() (which maybe should be renamed ReOrder() or something) and remove the duplicates with DeDup() and then output a new CSV file that looks the way I want it to.

Command syntax is dumb and arcane rn. Everything is hacky.

Want to add Filtering (only include rows where value in column matches something(s)) and bulk manipulation of per-column data (like add a string before and/or after the column data or change it or w/e)